+++ R/thicken_cust.R
@@ -13,6 +13,8 @@
 #' @param by Only needs to be specified when \code{x} contains multiple
 #' variables of class \code{Date}, \code{POSIXct} or \code{POSIXlt}.
 #' Indicates which to use for thickening.
+#' @param drop Should the original datetime variable be dropped from the
+#' returned data frame? Defaults to \code{FALSE}.
 #' @details
 #' Only rounding down is available for custom thickening.
 #' @return The data frame \code{x} with the variable added to it.
@@ -29,9 +31,11 @@
 thicken_cust <- function(x,
                          spanned,
                          colname,
-                         by = NULL) {
+                         by   = NULL,
+                         drop = FALSE) {
 
   is_df(x)
+  has_rows(x)
 
   original_data_frame <- x
   x <- as.data.frame(x)
@@ -39,11 +43,6 @@ thicken_cust <- function(x,
   dt_var_info <- get_dt_var_and_name(x, by)
   dt_var      <- dt_var_info$dt_var
   dt_var_name <- dt_var_info$dt_var_name
-
-  if (check_for_sorting(dt_var)){
-    warning('Datetime variable was unsorted, result will be unsorted as well.',
-            call. = FALSE)
-  }
 
   is_datetime(spanned)
   if (inherits(spanned, 'POSIXt') & inherits(dt_var, 'POSIXt')) {
@@ -61,12 +60,14 @@ thicken_cust <- function(x,
 
   dt_var <- check_for_NA_thicken(dt_var, dt_var_name, colname)
 
-  thickened <- round_thicken(dt_var, spanned, "down")
+  thickened <- round_thicken(dt_var, spanned, "down", ties_to_earlier = FALSE)
 
   thickened_frame <- data.frame(thickened)
 
   return_frame <- dplyr::bind_cols(x, thickened_frame)
   colnames(return_frame)[ncol(return_frame)] <- colname
+
+  if (drop) return_frame <- remove_original_var(return_frame, dt_var_name)
 
   set_to_original_type(return_frame, original_data_frame)
 }