+++ src/round_up_core_prev.cpp
@@ -2,26 +2,20 @@
 using namespace Rcpp;
 
 // [[Rcpp::export]]
-NumericVector round_down_core(NumericVector a, NumericVector b){
+IntegerVector round_up_core_prev(IntegerVector a, IntegerVector b){
 
   int n = a.size();
-  NumericVector ret(n);
+  IntegerVector ret(n);
   int current_b = b[0];
-  int next_b    = b[1];
   int current_b_index = 0;
-  int last_b    = max(b);
 
   for (int i = 0; i < n; ++i) {
-    if (next_b > a[i] or current_b == last_b) {
+    if (current_b >= a[i]) {
       ret[i] = current_b;
     } else {
-      while(next_b <= a[i]) {
+      while(current_b < a[i]) {
         current_b_index += 1;
         current_b = b[current_b_index];
-        if (next_b == last_b) {
-         break;
-        }
-        next_b    = b[current_b_index + 1];
       }
       ret[i] = current_b;
     }